==> code/9 - Creation of Voter Participation Data For Even and Odd Years.ipynb <==
##  We're going to load the turnout data, flatten it, and add it to key demographic data to create a comprehensive database table for query and predictions.  

### We will consider that a person has 'voted' if they voted in any election in an odd year and the November general election in an even year


#First set everthing up
import pyodbc
import graphlab as gl
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import datetime
import local #server information and credentials
from pullODBC import pullODBC
from pullODBCall import pullODBCall
from matplotlib.backends.backend_pdf import PdfPages
from __future__ import division # convert integer division to floating point
%matplotlib inline
plt.style.use('seaborn-white')
plt.rcParams['font.size'] = 7
plt.tight_layout;
plt.rc('xtick', labelsize=7) 
plt.rc('ytick', labelsize=7) 

### Section A: Extracting Voter History Data for each Registered Voter

Read in voter history from a csv file

history=pd.read_csv('history.csv')

Convert all columns to string

for col in history.columns:
    history[col]= history[col].apply(str)

Get rid of all the double quotes from the dataframe

history.columns = history.columns.str.strip('"')
for i, col in enumerate(history.columns):
    history.iloc[:, i] = history.iloc[:, i].str.replace('"', '')
print history.head(n=15)

Drop the junk column

history = history.drop(['Unnamed: 0'], axis=1)

Save the total number of rows to raw_len and print

raw_len=len(history)
print "The number of rows pulled from the database is " + str(raw_len)

Create a generic function to change blank fields to NaN for all categorical fields 

def fix_blanks_NaN(x):
    if x == "":
        return np.nan
    elif x == " ":
        return np.nan
    else:
        return x

history['election_lbl'] = history['election_lbl'].apply(fix_blanks_NaN) # change blanks to NaN
history['ncid'] = history['ncid'].apply(fix_blanks_NaN) # etc.
history['county_desc'] = history['county_desc'].apply(fix_blanks_NaN) # etc.

Clean the data by dropping rows with missing data and print a summary

history.dropna(inplace=True, how='any') #drop all rows with NaN
clean_len=len(history) #how many rows to I have left?
difference = raw_len - clean_len #how many rows have I dropped?
fraction = float(difference)/float(raw_len) #what fraction of rows have I dropped?
print "The number of rows I have dropped is " + str(difference) 
print "The number of rows remaining in the database is " + str(clean_len)
print "The precentage of rows that I have dropped is: " + str('{:.3%}'.format(fraction))

We need to flatten the database to a single row per NCID and massage the data

For local elections, election dates vary across the state in all odd numbered years.  In fact, many candidates run unopposed in the general election, so the primary election is more meaningfull.  Previously, we had only looked at elections held on election day in November.  For local elections, we will consider that a registered voter 'voted' if he voted in any election in odd numbered years.  So, the code below provides a marker of '1' if the voter voted at all in a year, '0' if that voter did not participate in any election that year.

In order to do this, we need to create a custom aggfunc that works on a pandas series and returns '1' if then length of that series is not equal to zero, and returns a '0' if the length of that series is equal to zero.

def not_zero_len(x):
    if len(x) != 0:
        return 1
    else:
        return 0

Let's create the new column 'year' that contains the year of the vote

history['year'] = history['election_lbl'].str[-4:]

print history.head()

Now we create a function that returns 'True' if the input number is even, 'False' if it is odd

def is_even(x):
    if (x%2 == 0): 
        return True
    else: 
        return False

Next, we create a new column containing 'True' for even years, 'False' for odd years

history['even'] = history['year'].astype(int).apply(is_even)

print history.head(n=50)

Now, let's split history into two dataframes, one for even years and one for odd years

history_even = history[history['even']]
history_odd = history[~history['even']]

print len(history_odd)/(len(history_odd)+len(history_even))

print history_odd.head(n=20)

del history # we no longer need the full 'history' dataframe

For even year elections, eliminate all records not in November

history_even = history_even[history_even['election_lbl'].astype(str).str[0:2]=='11']

print history_even.head()

print history_odd.head()

Now, let's use pandas pivot table technology to flatten the history file for even year elections

elections_by_ncid = pd.pivot_table(history_even[['ncid','year']], index=['ncid'],  
                                 aggfunc=len, columns='year', margins = False) # create a pivot table
elections_flattened_even = pd.DataFrame(elections_by_ncid.to_records()) # convert the pivot table to a dataframe
cols = elections_flattened_even.columns # get an index object with all the column names
cols = cols.delete(0) # remove the first column name 'ncid' since I don't want to rename this
elections_flattened_even.fillna(value=0, method=None, axis=None, inplace=True, limit=None, downcast=None) # replace NaN with 0
elections_flattened_even[cols] = elections_flattened_even[cols].astype(int)# convert values to integers
elections_flattened_even = elections_flattened_even.reindex_axis(sorted(elections_flattened_even.columns), axis=1) # sort the columns by year

print elections_flattened_even.head()

Now, let's use pandas pivot table technology to flatten the history file for odd year elections

First, we keep the aggfunc as 'len' so the output is the number of elections voted in each year

elections_by_ncid = pd.pivot_table(history_odd[['ncid','year']], index=['ncid'],  
                                 aggfunc=len, columns='year', margins = False) # create a pivot table
elections_flattened_odd = pd.DataFrame(elections_by_ncid.to_records()) # convert the pivot table to a dataframe
cols = elections_flattened_odd.columns # get an index object with all the column names
cols = cols.delete(0) # remove the first column name 'ncid' since I don't want to rename this
elections_flattened_odd.fillna(value=0, method=None, axis=None, inplace=True, limit=None, downcast=None) # replace NaN with 0
elections_flattened_odd[cols] = elections_flattened_odd[cols].astype(int)# convert values to integers
elections_flattened_odd = elections_flattened_odd.reindex_axis(sorted(elections_flattened_odd.columns), axis=1) # sort the columns by year

print elections_flattened_odd.head(n=10)

Next, we set the aggfunc to 'not_zero_len'  so the output is '1' if voted in any election that year, '0' if not voted in any elections that year

elections_by_ncid = pd.pivot_table(history_odd[['ncid','year']], index=['ncid'],  
                                 aggfunc=not_zero_len, columns='year', margins = False) # create a pivot table
elections_flattened_odd = pd.DataFrame(elections_by_ncid.to_records()) # convert the pivot table to a dataframe
cols = elections_flattened_odd.columns # get an index object with all the column names
cols = cols.delete(0) # remove the first column name 'ncid' since I don't want to rename this
elections_flattened_odd.fillna(value=0, method=None, axis=None, inplace=True, limit=None, downcast=None) # replace NaN with 0
elections_flattened_odd[cols] = elections_flattened_odd[cols].astype(int)# convert values to integers
elections_flattened_odd = elections_flattened_odd.reindex_axis(sorted(elections_flattened_odd.columns), axis=1) # sort the columns by year

print elections_flattened_odd.head(n=10)

Now, let's join the odd and even dataframes.  We use an outer join so we keep all NCIDs

elections_flattened = pd.merge(elections_flattened_odd, elections_flattened_even, how='outer')
print len(elections_flattened), len(elections_flattened_odd), len(elections_flattened_even)

print elections_flattened.head(n=10)

Just in case, change any 'NaN' values to zero

elections_flattened.fillna(value=0, method=None, axis=None, inplace=True, limit=None, downcast=None)

Finally, let's sort by columns in ascending order and get rid of '2018'

elections_flattened = elections_flattened.drop('2018', axis=1)
elections_flattened.sort(columns=None, axis=1, ascending=True, inplace=True, kind='quicksort', na_position='last')

print elections_flattened.head()

Clean up dataframes that are no longer necessary

del history_even, history_odd, elections_flattened_even, elections_flattened_odd

elections_flattened.to_csv('elections_flattened.csv')

We now have a flat dataframe with one row for each ncid and a column for each year.  The value in that column is '1' if the person with the ncid voted in the general election (even years) or any election (odd year) and '0' if they didn't vote

### Section B: Extracting Voter Demographics

Read in voter demographics from a csv file

demographics = pd.read_csv('demographics.csv')

Convert all columns to string

for col in demographics.columns:
    demographics[col]= demographics[col].apply(str)

Get rid of all the double quotes from the dataframe

demographics.columns = demographics.columns.str.strip('"')
for i, col in enumerate(demographics.columns):
    demographics.iloc[:, i] = demographics.iloc[:, i].str.replace('"', '')
print demographics.head()

#Save the total number of rows to raw_len and print. 
raw_len=len(demographics)
print "The number of rows pulled from the database is " + str(raw_len)

print demographics.columns

Create two lists:  one for fields that can have a value "None" and another for fields that must be deleted if blank

can_have_none = ['municipality_abbrv', 'birth_state', 'last_name', 'first_name']
cannot_have_none = ['birth_age', 'gender_code', 'race_code',
       'ethnic_code', 'party_cd', 'status_cd', 'zip_code', 
       'drivers_lic', 'county_desc', 'registr_dt', 'ncid',
       'precinct_abbrv']

Create a generic function to change blank fields to 'None' for 'birth_state', 'municipality_abbrv', 'first_name' and 'last_name'

def fix_blanks(x):
    if x == "":
        return "None"
    elif x == " ":
        return "None"
    else:
        return x

for col in can_have_none:
    demographics[col] = demographics[col].apply(fix_blanks)
    
for col in cannot_have_none:
    demographics[col] = demographics[col].apply(fix_blanks_NaN)

demographics.dropna(inplace=True, how='any') #drop all rows with NaN
clean_len=len(demographics) #how many rows to I have left?
difference = raw_len - clean_len #how many rows have I dropped?
fraction = float(difference)/float(raw_len) #what fraction of rows have I dropped?
print "The number of rows I have dropped is " + str(difference) 
print "The number of rows remaining in the database is " + str(clean_len)
print "The precentage of rows that I have dropped is: " + str('{:.3%}'.format(fraction))

### Section C: Merging Demographics and Turnout History

Now we do a left join, adding turnout information to each ncid in the demographics dataframe

combined = pd.merge(demographics, elections_flattened, how='left', on='ncid')

print len(combined), len(demographics), len(elections_flattened)

print combined.head()

Now save the dataframe 'combined' in a .csv

combined.to_csv('combined_updated.csv')
