==> code/9a - Modelling Voter Turnout for Local Elections.ipynb <==
##  We're going to load the combined demographic and turnout data, and build a model to predict turnout for the 2017 local elections - More Data - Do Not Change

#First set everthing up
import pyodbc
import graphlab as gl
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import datetime
import local #server information and credentials
from pullODBC import pullODBC
from pullODBCall import pullODBCall
from matplotlib.backends.backend_pdf import PdfPages
from __future__ import division # convert integer division to floating point
%matplotlib inline
plt.style.use('seaborn-white')
plt.rcParams['font.size'] = 7
plt.tight_layout;
plt.rc('xtick', labelsize=7) 
plt.rc('ytick', labelsize=7) 

Read in combined demographic and participation data from the June 2017 snapshots

combined_df = pd.read_csv('combined_updated.csv')

print len(combined_df)
print combined_df.head()
print combined_df.columns

Change all the NaN's to 0

combined_df.fillna(value=0.0, method=None, axis=None, inplace=True, limit=None, downcast=None)

pd.set_option('max_columns',100)

Get rid of junk columns

del combined_df['Unnamed: 0']
del combined_df['Unnamed: 0.1']

Change all particpation columns to integer

years = ['2007', '2008', '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016']
for year in years:
    combined_df[year] = combined_df[year].apply(int)

print combined_df.head()

We're only interested in municipal elections, so let's only keep rows where municiaplity is not equal to 'Blank"

print "The total number of registered voters is " + str(len(combined_df))
combined_df = combined_df[combined_df['municipality_abbrv']!= 'None']
print "The total number of registered voters living in municipalities is " +str(len(combined_df))
print combined_df.head()

Change registr_dt to datetime and birth_age to float

combined_df['birth_age'] = combined_df['birth_age'].apply(float)

combined_df['registr_dt']=pd.to_datetime(combined_df.registr_dt, format='%m/%d/%Y')
print combined_df.head()

Create datetime objects for the four election days to be used in the model

election_day_2015 = pd.to_datetime('11/3/2015', dayfirst=False, yearfirst=False)
election_day_2014 = pd.to_datetime('11/4/2014', dayfirst=False, yearfirst=False) 
election_day_2013 = pd.to_datetime('11/5/2013', dayfirst=False, yearfirst=False) 
election_day_2012 = pd.to_datetime('11/6/2012', dayfirst=False, yearfirst=False) 

Create a set of functions to set new columns to 'Y' (voted) 'N' (registered but did not vote) and 'U' (not registered)

def calc_voted_2015(x):
    if x['2015'] == 1:
        return 'Y'
    elif x['registr_dt'] <= election_day_2015:
        return 'N'
    else:
        return 'U'

def calc_voted_2014(x):
    if x['2014'] == 1:
        return 'Y'
    elif x['registr_dt'] <= election_day_2014:
        return 'N'
    else:
        return 'U'

def calc_voted_2013(x):
    if x['2013'] == 1:
        return 'Y'
    elif x['registr_dt'] <= election_day_2013:
        return 'N'
    else:
        return 'U'

def calc_voted_2012(x):
    if x['2012'] == 1:
        return 'Y'
    elif x['registr_dt'] <= election_day_2012:
        return 'N'
    else:
        return 'U'

combined_df['target']= combined_df.apply(calc_voted_2015, axis=1)
print "created target column"
combined_df['congressional']= combined_df.apply(calc_voted_2014, axis=1)
print "created congressional column"
combined_df['local']= combined_df.apply(calc_voted_2013, axis=1)
print "created local column"
combined_df['presidential']= combined_df.apply(calc_voted_2012, axis=1)
print "created presiendial column"

Print the first 10 rows and examine for correctness

pd.set_option('max_columns', 30)
print combined_df.head(n=10)

Let's clean up the dataframe by only keeping the columns that we need and print the top to inspect and delete the old dataframe that we no longer need

participation_df = combined_df[['ncid', 'birth_age', 'gender_code', 'race_code', 'ethnic_code', 'party_cd', 
                               'status_cd', 'zip_code','birth_state', 'drivers_lic', 'county_desc', 'registr_dt',
                               'precinct_abbrv','target', 'congressional', 'local', 'presidential', 'municipality_abbrv']]

del combined_df

print participation_df.head()

Before modelling, need to make sure the columns are of the correct type and that all folks with invalid ages and years registered are removed.  SFrames, which will be used for modelling, do not accept mixed types in columns, so we need to be very sure that all column types are the same.

string_columns = ['ncid', 'gender_code',  'race_code',  'ethnic_code', 'party_cd', 'status_cd', 
                'zip_code', 'birth_state', 'drivers_lic', 'county_desc', 'precinct_abbrv',
                 'target', 'congressional', 'local', 'presidential', 'municipality_abbrv']

for col in string_columns: #convert all the string_columns to 'string'
    participation_df[col] = participation_df[col].apply(str)

Create interaction variables

def convert_to_years(x): # function to convert a timedelta object to a float of years
    return x.total_seconds()/(60*60*24*365.2422)

participation_df['race_gender'] = participation_df['race_code'] + participation_df ['gender_code']
participation_df['race_ethnicity'] = participation_df['race_code'] + participation_df['ethnic_code']
participation_df['gender_ethnicity'] = participation_df['gender_code'] + participation_df ['ethnic_code']
participation_df['age_squared'] = (participation_df['birth_age'])**2
participation_df['age_cubed'] = (participation_df['birth_age'])**3
now = pd.to_datetime('today') # get today's data
participation_df['years_since_registration'] = now - participation_df['registr_dt'] # get days since registration as a timedelta object
participation_df['years_since_registration'] = participation_df['years_since_registration'].apply(convert_to_years) # convert to years
participation_df['years_squared'] = (participation_df['years_since_registration'])**2
participation_df['years_cubed'] = (participation_df['years_since_registration'])**3

print participation_df.head()

Remove all of the voters with bad ages and registration dates

participation_df = participation_df[(participation_df['birth_age'] <= 99) & (participation_df['years_since_registration'] <= 61)]

To create the model, we can only use voters that were registered in 2015.  So we need to remove all voters with code 'U' in the 'target' column.  

model_df = participation_df[(participation_df['target']!= 'U')]

Let's figure out the majority classifier

majority = model_df.groupby(['target'])['target'].count()/model_df.groupby(['target'])['target'].count().sum()
print majority
size_all = len(model_df)
print size_all

By definition of status_cd = 'I', this implies that voters are inactive.  Therefore, the logical thing for our model to do is to predict all voters with status_cd = 'I' will not vote in the next local election.  So, we should model only the voters with status_cd = 'A'.  We can then combine everything in our final classification model to predict whether a voter will vote in the next municipal election:

    if status_cd == 'I' then predict 'No'

        else if model_prediction == 'Y' then predict 'Yes'
    
        else predict 'No'

Let's remove all voters with status_cd = 'I' from the model dataframe

model_df = model_df[(model_df['status_cd']== 'A')]

majority = model_df.groupby(['target'])['target'].count()/model_df.groupby(['target'])['target'].count().sum()
print majority
size_active = len(model_df)
print "The number of both active and inactive voters is " +str(size_all)
print "The number of active voters is " + str(size_active)
print str(size_all - size_active) + " voters are inactive."
print "The fraction of voters that are inactive is " +str((size_all-size_active)/size_all)

Our majority classifier is 83.1%.  This could be our starting benchmark for accuracy, since if we guess that everyone registered for the 2015 election did not vote, we would guess correctly 83.1% of the time.  

However, since the classes are imbalanced this is not a good benchmark.  What we're really trying to do is identify which voters are likely to vote in the next local election.  If we use the majority classifier, we would predict that no one would vote.

Accuracy of correclty predicting voters that will vote: 0%

Accuracy of correctly predicting voters that will not vote: 100%

Therefore, we're going to try to balance accuracy of predicting both voters that will vote and voters that won't vote.  

We can now explore 2015 turnout by county/muncipality for active voters only

table = pd.pivot_table(model_df[['county_desc', 'municipality_abbrv', 'target']], values=None, index=['county_desc','municipality_abbrv'], 
                                  columns='target', aggfunc=len, fill_value=None, margins=False, dropna=True, margins_name='All')
table = pd.DataFrame(table.to_records()) # convert the pivot table to a dataframe
table['turnout_active'] = table['Y']/(table['Y']+table['N'])

print table

Now we convert model_df to an SFrame for modelling.

model_sf = gl.SFrame(model_df)

cols = model_sf.column_names()

print cols

Let's start out by using only past participation as our features

features = ['congressional', 'local', 'presidential']                

print features

We now reduce the data we will model to include only the features plus the column 'target'

model_reduced = model_sf[['congressional', 'local', 'presidential', 'target']]

Split the model data into 80% trainng data and 20% test data

train_data,test_data = model_reduced.random_split(.8)

In order to yeild meaningful coefficients, I need to insert a top row onto the SFrame with 'U' as values.

reference_sf = gl.SFrame({'local':'U', # create a new SFrame with 1 row and "None" for all categorical variables
                    'target': 'N', 
                    'congressional': 'U', 
                    'presidential': 'U'})
train_data = reference_sf.append(train_data) # append the full SFrame to the bottom of this 1 row SFrame 
print train_data.head()

We will no train and test a logistic classifier model.  To improve performance, we will use a Fast Iterative Shrinkage-Thresholding Algorithm (FISTA).  This algorithm speeds up convergence by 

turnout_model = gl.logistic_classifier.create(train_data, target='target',features= features,validation_set=test_data,
                                                          verbose=True, solver='newton', class_weights=None, max_iterations=50)

turnout_model.evaluate(test_data)

Accuracy is 85.1% for using only previous turnout as features vs. 83.1% for the majority classifier, only a small improvement.   However, this does not tell the whole story.  Recall that the majority classifier performed as fullows:

Accuracy of correclty predicting voters that will vote: 0%

Accuracy of correctly predicting voters that will not vote: 100%


The classifier model we've just created performs as follows:

Accuracy of correclty predicting voters that will vote: 58%

Accuracy of correctly predicting voters that will not vote: 90%

If our objective to predict voters that actually will vote, this model is far better than the majority classifier.  However, identifying voters with 58% accuracy is still not good enough. 

We've observed that local election turnout varies widely among counties.  Let's add county_desc to the model and see if there is any meaningful improvement.

enhanced_features = ['congressional', 'local', 'presidential',  'county_desc'] 

model_reduced = model_sf[['local', 'target', 'congressional', 'presidential', 'county_desc']]
train_data,test_data = model_reduced.random_split(.8)

reference_sf = gl.SFrame({'local':'U', # create a new SFrame with 1 row and "None" for all categorical variables
                    'target': 'N', 
                    'congressional': 'U', 
                    'presidential': 'U',
                    'county_desc': 'U'})
train_data = reference_sf.append(train_data) # append the full SFrame to the bottom of this 1 row SFrame 
print train_data.head()

turnout_model = gl.logistic_classifier.create(train_data, target='target',features= enhanced_features,validation_set=test_data,
                                                          verbose=True, solver='newton', class_weights=None, max_iterations=100)

turnout_model.evaluate(test_data)

Accuracy is 86.7% for using  previous turnout + county as features vs. 85.1 % using only previous turnout, again a small improvement.   However, this does not tell the whole story.  Recall that the classifier using only previous turnout performed as follows:

Accuracy of correctly predicting voters that will vote: 58%

Accuracy of correctly predicting voters that will not vote: 90%

Our new classifier adding county to the feature list performed as follows:

Accuracy of correctly predicting voters that will vote: 43%

Accuracy of correctly predicting voters that will not vote: 95%

If our objective to predict voters that actually will vote, this goes in the wrong direction.  The new model became more accurate at predicting non-voters at the expense of the accuracy of predicting voters.  We need to fix this.

Fortunately, there is a simple way do resolve this problem.  To achieve the best accuracy, the algorithm as it is considers the penalty for predicting a non-voter incorrectly as equal to the penalty for predicting a voter incorrectly.  This does not match our real-world objectives.  Classification modeling using unbalanced classes is a well studied subject, and one recommended way to handle this situation is to penalize an incorrect prediction into the majority class more than an incorrect prediction into the minority class.  This will increase the accuracy of predicting voters at the expense of the accuracy of predicting non-voter.  Overall accuracy will be lower, but this is a fair price to pay for increased accuracy of voters.  One rule of thumb is penalize incorrect prediction into the majority class proportionally to the class imbalance.  This would imply a penalty of 83.1/16.9.  In other words, the penalty for incorrectly predicting a voter as a non-voter would be approximately 5 times that of incorrectly predicting a voter as a non-voter.  Let's try modeling this way and see if we can get a better result.

turnout_model = gl.logistic_classifier.create(train_data, target='target',features= enhanced_features,validation_set=test_data,
                                                          verbose=True, solver='newton', class_weights='auto', max_iterations=100)

turnout_model.evaluate(test_data)

Accuracy is 80.6% for using  previous turnout + county as features vs. 86.6% using the unweighted model, significantly lower.  Once again, this does not tell the whole story.  Recall that the classifier using the unweighted model performed as follows:

Accuracy of correctly predicting voters that will vote: 43%

Accuracy of correctly predicting voters that will not vote: 95%

Using the weighted model, performance was:

Accuracy of correctly predicting voters that will vote: 73%

Accuracy of correctly predicting voters that will not vote: 82%


If our objective to predict voters that actually will vote, this is a significant improvement.  The new model became more accurate at predicting voters at the expense of the accuracy of predicting non-voters.  This is exactly what we needed.

To further improve accuracy of correctly predicting voters (make it as close as possible to the accuracy of correctly predicting non-voters), We can adjust the threshold for predicting non-voters lower.  Currently, every person with a probability score > 0.5 will be predicted as a voter, and those < 0.5 will be predicted as a non-voter.  Let's move this threshold down to balance the accuracies.

We now need to create a function that manually calculate accuracy, sensitivity and specificity.

The function below (get_model_performance) takes as input a classification model, an SFrame used by this model to make predictions, and a threshold for making predictions such that any probability above this threshold yields s 'positive' or 'republican' classification while any probability below this threshold yields a 'negative' or 'democratic' classification. It also takes as input the column in the SFrame containing the classification target and the value of this column that represents a 'positive'.  As a reminder, we are considering 'republican' sentiment as a positive and 'democratic' sentiment as a negative.

This function returns a 3-tuple:  accuracy, sensitivity, selectivity

from __future__ import division # convert integer division to floating point
def get_model_performance(model, sframe, threshold, target_feature, positive_value):
    sframe['percent'] = model.predict(sframe, output_type='probability') # add a column with the model generated probabilities
    sframe['predict_postive'] = sframe['percent'] >= threshold # add a column that contains 'true' if a potivite is predicted
    sframe['predict_negative'] = sframe['percent'] < threshold # add a column that contains 'true' if a negative is predicted
    sframe['true_postive'] = sframe[target_feature] == positive_value # add a column that contains 'true' if the row is a true postive
    sframe['true_negative'] = sframe[target_feature] != positive_value # add a column that contains 'true' if the row is a true negative
    
    num_observations = len(sframe) # total number of observations
    true_positives = sframe['true_postive'].sum() # number of true postives
    true_negatives = sframe['true_negative'].sum() # number of true negatives
    correct_positives = ((sframe['true_postive'] == 1) & (sframe['predict_postive'] == 1)).sum() # number of correct positive predictions
    correct_negatives = ((sframe['true_negative'] == 1) & (sframe['predict_negative'] == 1)).sum() # number of correct negative predictions
    sensitivity = correct_positives/true_positives # calculate sensitivity: correctly predicted postives / total positives
    specificity = correct_negatives/true_negatives  # calculate specificity: correclty predicted negatives / total negatives
    accuracy = (correct_positives + correct_negatives)/num_observations # calculate accuracy:  correct predictions / total observations
    return accuracy, sensitivity, specificity # return the 3-tuple

Let's test the above function

print get_model_performance(turnout_model, test_data, .5, 'target', 'Y')

Let's use the above function to select the threshold that balances the accuracies.

min_difference = 1
best_threshold = 0
best_result = 0.0,0.0,0.0
for thresh in np.linspace(0.45, 0.5, num=100):
    output =  get_model_performance(turnout_model, train_data, thresh, 'target', 'Y')
    difference = output[1]-output[2]
    if abs(difference) < min_difference:
        min_difference = abs(difference)
        best_threshold = thresh
        best_result = output
print "The best threshold value for predicting positives is " + str(best_threshold)  
print "This yields accuracy of " + str(best_result[0]) + " sensitivity of " + str(best_result[1]) + " and specificity of " + str(best_result[2])
threshold_to_use = best_threshold  

We now have a good model.  To summarize:

Accuracy of correctly predicting voters that will vote: 79%

Accuracy of correctly predicting voters that will not vote: 77%

If model probability score < 0.488 then predict "Will not vote"

Else (probability score > 0.488) predict "Will vote"

This model can be combined with 'status_cd' = 'I' ----> "Will not vote"  to make final predictions.

Let's explore the coefficients to determine the influence each has on the model.

coeff_sf = turnout_model.coefficients
print coeff_sf.print_rows(num_rows=200)

By far, the factor that has the most impact on projecting turnout for an upcoming local election is whether a person voted in the last local election.  The next most meaningful factor is whether a person voted in the last congressional election.  Whether a person voted in the last presidential election has little impact on projecting turnout for the next local election.  

Which county you live in in North Carolina has a large impact in whether you will vote in a local election.  That being said, county is very highly correlated with voting history.  This high correlation is demonstrated by the very small (1.6%) increase in accuracy we achieve by adding county to voting history.

Let's explore the coefficients in more detail

turnout_model.save('turnout_model')

coeff_sf.save('turnout_coeff_sf')

Create a new dataframe for coefficient statistics, calculate these statistics, and add these to the dataframe.  First, create a pandas dataframe from the coefficient sframe.

coeff_df = coeff_sf.to_dataframe()

import statsmodels.robust.scale as sm # for median absolute deviation (MAD)

coeff_df['value'].astype('float64') # Ensure that values are float64
coeff_statistics_df = pd.DataFrame(data=None, index=None, columns=['COEF','MIN', 'MAX', 'RANGE', 'MAD'], 
                                 dtype='float64', copy=False) # Create an empty dataframe for statistics
names = coeff_df['name'].unique() # names has all the feature values 
# Iterate through all the features with more than 1 coeffient value (the categorical varaiabes) 
# and add statistics about the coefficient values to coeff_statistics_df
for coeff in names:
    df = coeff_df[coeff_df['name']== coeff]
    if len(df) > 1:
        new_row = pd.DataFrame([[coeff, df['value'].min(), df['value'].max(), df['value'].ptp(), 
                                 sm.mad(df['value'])]], columns=['COEF','MIN', 'MAX', 'RANGE', 'MAD']) 
                               
        coeff_statistics_df = coeff_statistics_df.append(new_row, ignore_index=False)
        
coeff_statistics_df.sort_values('MAD', axis=0, ascending=False, inplace=True, 
                              kind='quicksort', na_position='last') # sort the dataframe on Median Absolute Deviation

print coeff_statistics_df

The table above supports our previous conclusions

Let's try using a combination of county and precicnt_desc as our geographical predictor instead of just county. We create the new column first.

model_sf['county_prec'] = model_sf['county_desc'] + model_sf['precinct_abbrv']

Next we create the feature list

new_enhanced_features = ['congressional', 'local', 'presidential',  'county_prec'] 

model_reduced = model_sf[['local', 'target', 'congressional', 'presidential', 'county_prec']]
train_data,test_data = model_reduced.random_split(.8)

reference_sf = gl.SFrame({'local':'U', # create a new SFrame with 1 row and "None" for all categorical variables
                    'target': 'N', 
                    'congressional': 'U', 
                    'presidential': 'U',
                    'county_prec': 'U'})
train_data = reference_sf.append(train_data) # append the full SFrame to the bottom of this 1 row SFrame 
print train_data.head()

turnout_model = gl.logistic_classifier.create(train_data, target='target',features= new_enhanced_features, 
                                              validation_set=test_data, verbose=True, solver='newton', 
                                              class_weights='auto', max_iterations=100)

turnout_model.evaluate(test_data)

Accuracy is 80.8% for using  previous turnout + county as features vs. 80.6% using the county only model, increased slightly.  Recall that the classifier using only county performed as follows:

Accuracy of correctly predicting voters that will vote: 73%

Accuracy of correctly predicting voters that will not vote: 82%

The new model using county + precinct performed as follows:

Accuracy of correctly predicting voters that will vote: 80%

Accuracy of correctly predicting voters that will not vote: 81%


If our objective to predict voters that actually will vote, this is a significant improvement.  The new model became more accurate at predicting voters at the slight expense of the accuracy of predicting non-voters.  This is exactly what we needed.  These numbers are so close that tuning might be of limited value but we'll try it anyway to see if we can slightly improve the model


min_difference = 1
best_threshold = 0
best_result = 0.0,0.0,0.0
for thresh in np.linspace(0.47, 0.5, num=100):
    output =  get_model_performance(turnout_model, train_data, thresh, 'target', 'Y')
    difference = output[1]-output[2]
    if abs(difference) < min_difference:
        min_difference = abs(difference)
        best_threshold = thresh
        best_result = output
print "The best threshold value for predicting positives is " + str(best_threshold)  
print "This yields accuracy of " + str(best_result[0]) + " sensitivity of " + str(best_result[1]) + " and specificity of " + str(best_result[2])
threshold_to_use = best_threshold  

Wow!  It looks like we hit the mark.  We'll use this new model to predict turnout for 2017 municipal elections by registered voter.  

coeff_sf = turnout_model.coefficients

turnout_model.save('enhanced_turnout_model')

coeff_sf.save('enhanced_turnout_coeff_sf')

Create a new dataframe for coefficient statistics, calculate these statistics, and add these to the dataframe.  First, create a pandas dataframe from the coefficient sframe.

coeff_df = coeff_sf.to_dataframe()

import statsmodels.robust.scale as sm # for median absolute deviation (MAD)

coeff_df['value'].astype('float64') # Ensure that values are float64
coeff_statistics_df = pd.DataFrame(data=None, index=None, columns=['COEF','MIN', 'MAX', 'RANGE', 'MAD'], 
                                 dtype='float64', copy=False) # Create an empty dataframe for statistics
names = coeff_df['name'].unique() # names has all the feature values 
# Iterate through all the features with more than 1 coeffient value (the categorical varaiabes) 
# and add statistics about the coefficient values to coeff_statistics_df
for coeff in names:
    df = coeff_df[coeff_df['name']== coeff]
    if len(df) > 1:
        new_row = pd.DataFrame([[coeff, df['value'].min(), df['value'].max(), df['value'].ptp(), 
                                 sm.mad(df['value'])]], columns=['COEF','MIN', 'MAX', 'RANGE', 'MAD']) 
                               
        coeff_statistics_df = coeff_statistics_df.append(new_row, ignore_index=False)
        
coeff_statistics_df.sort_values('MAD', axis=0, ascending=False, inplace=True, 
                              kind='quicksort', na_position='last') # sort the dataframe on Median Absolute Deviation

print coeff_statistics_df

Let's compare the above to the previous model.  The MAD for local, congressional, and presidential previous participation is pretty much unchanged.  The MAD for county_prec is significantly higher than county alone.  This tells us that adding precinct to the county give us meaningful information as to turnout probability.  That being said, we need to be careful; county_prec is still highly correlated with the turnout information so we need to use this information carefully.  In other words, in specific precincts, there may be no real person that actually voted in the last 3 elections so if we were to model hypothetical people assuming 100% turnout in these last 3 elections our predictions would probably be off.  

One final thought.  Municipal elections are very different from each other.  Some elections have no competitive races most years.  Other elections may be competitive some years and not competitive other years.  Statewide analyses and predictions are of limited use.  To most accurately model local elections, modeling needs to done for each unique municipality considering the specific circumstances of each election. I would suggest these be done a county at a time for each municipality in a county.  These models could be combined with the political preference model to help identify specific voters that have specific turnout probabilities combined with specific political leanings.  These voters could be contacted for focus groups and surveys for use in journalistic projects.
