==> code/9d - Exploring Data Analmolies.ipynb <==
##  We're going to look for data anomalies

#First set everthing up
import pyodbc
import graphlab as gl
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import datetime
import local #server information and credentials
from pullODBC import pullODBC
from pullODBCall import pullODBCall
from matplotlib.backends.backend_pdf import PdfPages
from __future__ import division # convert integer division to floating point
%matplotlib inline
plt.style.use('seaborn-white')
plt.rcParams['font.size'] = 7
plt.tight_layout;
plt.rc('xtick', labelsize=7) 
plt.rc('ytick', labelsize=7) 
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))

### Section A: Extracting Voter History Data for each Registered Voter  to find data anomalies

#Set up the fields that I need to pull from the database 
fields = '["ncid"], ["election_lbl"], ["county_desc"], ["pct_label"],["voted_county_desc"], ["voting_method"]'
print fields

Change the database to 'history' and the table to 'ncvhis_Statewide'

local.database = 'history'
local.table = 'ncvhis_Statewide'

history=pullODBCall(local.driver,
                      local.server,
                      local.database,
                      local.username,
                      local.password,
                      local.table,
                      fields)

Get rid of all the double quotes from the dataframe

history.columns = history.columns.str.strip('"')
for i, col in enumerate(history.columns):
    history.iloc[:, i] = history.iloc[:, i].str.replace('"', '')
print history.head(n=15)

#Save the total number of rows to raw_len and print. 
raw_len=len(history)
print "The number of rows pulled from the database is " + str(raw_len)

Create a generic function to change blank fields to NaN for all categorical fields 

def fix_blanks_NaN(x):
    if x == "":
        return np.nan
    elif x == " ":
        return np.nan
    else:
        return x

history['election_lbl'] = history['election_lbl'].apply(fix_blanks_NaN) # change blanks to NaN
history['ncid'] = history['ncid'].apply(fix_blanks_NaN) # etc.
history['county_desc'] = history['county_desc'].apply(fix_blanks_NaN) # etc.

Clean the data by dropping rows with missing data and print a summary

history.dropna(inplace=True, how='any') #drop all rows with NaN
clean_len=len(history) #how many rows to I have left?
difference = raw_len - clean_len #how many rows have I dropped?
fraction = float(difference)/float(raw_len) #what fraction of rows have I dropped?
print "The number of rows I have dropped is " + str(difference) 
print "The number of rows remaining in the database is " + str(clean_len)
print "The precentage of rows that I have dropped is: " + str('{:.3%}'.format(fraction))

Now we need to discard all elections not in November, since we only want to analyze general elections

history = history[history['election_lbl'].astype(str).str[0:2]=='11']

We need to flatten the database to a single row per NCID and massage the data

elections_by_ncid = pd.pivot_table(history[['ncid','election_lbl']], index=['ncid'],  
                                 aggfunc=len, columns='election_lbl', margins = False) # create a pivot table
elections_flattened = pd.DataFrame(elections_by_ncid.to_records()) # convert the pivot table to a dataframe
cols = elections_flattened.columns # get an index object with all the column names
cols = cols.delete(0) # remove the first column name 'ncid' since I don't want to rename this
elections_flattened.fillna(value=0, method=None, axis=None, inplace=True, limit=None, downcast=None) # replace NaN with 0
elections_flattened[cols] = elections_flattened[cols].astype(int)# convert values to integers
for col in cols: # rename each election column to only the election year rather than the full date
    elections_flattened.rename(index=str, inplace=True, columns={col : col[-4:]})
elections_flattened = elections_flattened.reindex_axis(sorted(elections_flattened.columns), axis=1) # sort the columns by year

print elections_flattened.head()

Let's examine some anomolies in the voter history data.  Notice above that there is a column for 2018.  This means there is a record in the data for 2018.  Let's first find out which NCID is assocated with this record.

del elections_by_ncid # This is not used anymore sol let's delete it to free up some memory
pd.set_option('display.height', 5000)
pd.set_option('display.max_rows', 5000)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)
print elections_flattened[elections_flattened['2018']==1]

We've identified NCID CW1008993.  Let's look at the entire November voting history for this NCID from 2007 thru 2016.

voter_2018 = history[history['ncid']== 'CW1008993']

print voter_2018

Apparently, the person associated with NCID CW1008993 only has a single record for 2018.  Since November 6th was election day in both 2007 and 2012, this record may be a typo from that year.

Let's now look at instances where more than one voting record exists for an ncid for a November election.  First, we see how many NCIDs are associated with multiple voting records for each year

for year in ['2007', '2008', '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016']:
        print len(elections_flattened[elections_flattened[year]>1])

Considering the fact that we have over 32M voting records, this is a relatively small number.  However, these cases need investigation.  Since the voter history file has over 32M records, it's difficult to reshape and analyze in it's raw form.  Therefore, we need to create a subset of the voter history file for analysis.  To do this, we'll first identify any and all ncids that have more than one voting record for a specific November election.  We will then create a subset of the voter histroy file only containing records with these ncids.

Lets group the data by NCID (only for NCIDs associated with multiple voting records) and generate a report for all of these NCIDs.

multiple_records = elections_flattened[(elections_flattened['2007']>1) | (elections_flattened['2008']>1)\
|(elections_flattened['2009']>1) | (elections_flattened['2010']>1)\
|(elections_flattened['2011']>1) | (elections_flattened['2012']>1)\
|(elections_flattened['2013']>1) | (elections_flattened['2014']>1)\
|(elections_flattened['2015']>1) | (elections_flattened['2016']>1)]

print len(multiple_records)

It looks like a total of 825 NCIDs have multiple records for November elections from 2007 thru 2016.  We need to select records from the voter history data corresponding only to these NCIDs.  First create an iterable series of NCIDs from multiple_records.

multiple_series = multiple_records['ncid']

Next, let's subset out the history dataframe to include only those records associated with voters having multiple records in any year

history_subset = history[history['ncid'].isin(multiple_series)]
print len(history_subset)

We have 4149 records.  Now let's group by NICD and election_lbl

### First stip out blanks from election_lbl
history_subset['election_lbl']= history_subset['election_lbl'].map(lambda x: x.strip())
history_subset = history_subset.sort_values(by = ['ncid', 'election_lbl'], ascending=[1, 1])
report = history_subset.set_index(['ncid', 'election_lbl', 'county_desc'])
print len(report)
print report.head(n=15)

This list is too long to analyze.  Let's only select the nicd / election_lbl combinations where there is more than one "vote"

# First, reset the indexs and create a groupby object
reset_report = report.reset_index()# Reset the index
num_voting_records = reset_report.groupby(['ncid', 'election_lbl'])['county_desc'].count() # Create a groupBy object with counts of number of "votes"
multiple = num_voting_records[num_voting_records >1] # Only keep entries with more than 1 "vote"

The series "multiple" contains all the indexes ( ncid / election_lbl combinations) where there was more than 1 "vote" recored.  Let's use this series to iterate on the "reset_report" dataframe to only select instances of more than one vote

index_list = list() # Create an empty list for the indexes to keep

for row in multiple.index.values:
    ncid = row[0]
    election_lbl = row[1]
    selected_row_indices = reset_report[(reset_report['ncid']== ncid) & (reset_report['election_lbl']== election_lbl)].index[0:2]
    index_list.append(selected_row_indices)
    

selected = [element for tupl in index_list for element in tupl]
new_report = reset_report[reset_report.index.isin(selected)]
#To simplify report, let's delete county_desc since this is only the current county where the voter is registered.  We will only focus on voted_county_desc
new_report.drop('county_desc',axis=1, inplace=True)
print len(new_report)/2
new_report =  new_report.set_index(['ncid', 'election_lbl', 'voted_county_desc'])

Let me know save this to a text file

%store new_report > multiple_voting_records_june2017_snapshot.txt

Clean up old dataframes

del history
del elections_flattened

### Section B: Extracting Voter Demographics to find data anomalies

First change the database and table names back to get registration data

local.database = 'registration'
local.table = 'ncvoter_Statewide'

We now do the standard extraction and cleaning.  I've excluded comments since task has been done many times

#Set up the fields that I need to pull from the database 
fields = '["birth_age"], ["gender_code"], ["race_code"], ["ethnic_code"],["party_cd"], ["status_cd"], ["zip_code"], ["birth_state"], ["drivers_lic"], ["county_desc"], ["registr_dt"],["ncid"],["precinct_abbrv"]'
print fields

condition1 = '["status_cd"] = ' + local.single_double_quote + 'A' + local.double_single_quote
condition2 = '["status_cd"] = ' + local.single_double_quote + 'I' + local.double_single_quote
condition3 = '["county_desc"] = ' + local.single_double_quote + 'Chatham' + local.double_single_quote
condition = "(" + condition1 + " OR " + condition2 + ")"  # This line will be run on Azure
#condition = "(" + condition1 + " OR " + condition2 + ")" + " AND " + condition3# This line will be run on local
print condition

demographics=pullODBC(local.driver,
                      local.server,
                      local.database,
                      local.username,
                      local.password,
                      local.table,
                      fields,
                      condition)

Get rid of all the double quotes from the dataframe

demographics.columns = demographics.columns.str.strip('"')
for i, col in enumerate(demographics.columns):
    demographics.iloc[:, i] = demographics.iloc[:, i].str.replace('"', '')
print demographics.head()

#Save the total number of rows to raw_len and print. 
raw_len=len(demographics)
print "The number of rows pulled from the database is " + str(raw_len)

Create a generic function to change blank fields to 'None' for some categorical fields 

def fix_blanks(x):
    if x == "":
        return "None"
    elif x == " ":
        return "None"
    else:
        return x

Create a generic function to change blank fields to NaN for all categorical fields 

def fix_blanks_NaN(x):
    if x == "":
        return np.nan
    elif x == " ":
        return np.nan
    else:
        return x

demographics['birth_state'] = demographics['birth_state'].apply(fix_blanks) # change blanks to "None" in birth_state
demographics['gender_code'] = demographics['gender_code'].apply(fix_blanks_NaN) # change blanks to NaN in gender_code
demographics['race_code'] = demographics['race_code'].apply(fix_blanks_NaN) # etc.
demographics['ethnic_code'] = demographics['ethnic_code'].apply(fix_blanks_NaN) 
demographics['zip_code'] = demographics['zip_code'].apply(fix_blanks_NaN)
demographics['birth_state'] = demographics['birth_state'].apply(fix_blanks_NaN) 
demographics['drivers_lic'] = demographics['drivers_lic'].apply(fix_blanks_NaN) 
demographics['county_desc'] = demographics['county_desc'].apply(fix_blanks_NaN) 
demographics['registr_dt'] = demographics['registr_dt'].apply(fix_blanks_NaN)
demographics['ncid'] = demographics['ncid'].apply(fix_blanks_NaN)
demographics['precinct_abbrv'] = demographics['precinct_abbrv'].apply(fix_blanks_NaN)

demographics.dropna(inplace=True, how='any') #drop all rows with NaN
clean_len=len(demographics) #how many rows to I have left?
difference = raw_len - clean_len #how many rows have I dropped?
fraction = float(difference)/float(raw_len) #what fraction of rows have I dropped?
print "The number of rows I have dropped is " + str(difference) 
print "The number of rows remaining in the database is " + str(clean_len)
print "The precentage of rows that I have dropped is: " + str('{:.3%}'.format(fraction))

def convert_to_float(x): # function to convert a timedelta object to a float of years
    return x.total_seconds()/(60*60*24*365.2422)

demographics['birth_age'] = demographics['birth_age'].astype('float64')
demographics['registr_dt'] = pd.to_datetime(demographics['registr_dt'], format='%m/%d/%Y')
demographics['years_since_registration'] = pd.to_datetime('today') - demographics['registr_dt'] # get days since registration as a timedelta object
demographics['years_since_registration'] = demographics['years_since_registration'].apply(convert_to_float) # convert to years in floating point


Frist, let's look at voters over 115 years of age

print len(demographics[demographics['birth_age']> 115])
print demographics[demographics['birth_age']> 115][['ncid', 'birth_age']].head(n=20)

These are obviously errors.  Let's create a report sorting these folks by county and storing this in a report.

subset = demographics[demographics['birth_age']> 115][['county_desc', 'ncid', 'birth_age','registr_dt']]
subset = subset.sort_values(by = ['county_desc','birth_age'], ascending=True)
subset = subset.set_index(['county_desc','birth_age']) #sort by county
%store subset > age_errors_june2017_snapshot.txt

Let's now create a report with counts by county

# First, reset the indexs and create a groupby object
subset = subset.reset_index()# Reset the index
num_invalid_age = subset.groupby(['county_desc'])['county_desc'].count() # Create a groupBy object with counts of number of voters with birth_age > 115
num_invalid_age.nlargest(100)# Sort by number of voters

It looks like six counties make up the bulk of the problem records: GUILFORD, FORSYTH, CUMBERLAND, RANDOLPH, HENDERSON, and HALIFAX      

Next, let's look at voters that have been registered for more than 100 years

print len(demographics[demographics['years_since_registration']> 100])
print demographics[demographics['years_since_registration']> 100][['ncid', 'registr_dt']].head(n=15)

These are obviously errors.  Let's create a report sorting these folks by county and storing this in a report.

subset = demographics[demographics['years_since_registration']> 100][['county_desc', 'ncid', 'registr_dt', 'birth_age']]
subset = subset.sort_values(by = ['county_desc','registr_dt'], ascending=True)
subset = subset.set_index(['county_desc','registr_dt']) #sort by county
%store subset > registration_errors_june2017_snapshot.txt

Let's now create a report with counts by county

# First, reset the indexs and create a groupby object
subset = subset.reset_index()# Reset the index
num_invalid_registration = subset.groupby(['county_desc'])['county_desc'].count() # Create a groupBy object with counts of number of voters with invalid registration dates
num_invalid_registration.nlargest(100)# Sort by number of voters

It looks like six counties make up the bulk of the problem records: CUMBERLAND, MCDOWELL, PERSON, ALAMANCE, DUPLIN, and HALIFAX
