==> code/9f - Calculating Turnout for Clustering.ipynb <==
##  We're going calculate turnout by county/precinct combination to use as 3 dimensions for our clustering work

#First set everthing up
import pyodbc
import graphlab as gl
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import datetime
import local #server information and credentials
from pullODBC import pullODBC
from pullODBCall import pullODBCall
from matplotlib.backends.backend_pdf import PdfPages
from __future__ import division # convert integer division to floating point
%matplotlib inline
plt.style.use('seaborn-white')
plt.rcParams['font.size'] = 7
plt.tight_layout;
plt.rc('xtick', labelsize=7) 
plt.rc('ytick', labelsize=7) 

Read in the full voter history

elections_flattened = pd.read_csv('elections_flattened.csv')

print len(elections_flattened)
print elections_flattened.head()
print elections_flattened.columns

Since we know that muliple voting records are present in some elections, we need to see if any number is > 1.  Note that we already eliminated these for odd years, so we should only see these for even years

### First we set up a list of column names for the years
year_columns = ['2007', '2008', '2009', '2010', '2011', '2012', '2013', '2015', '2015', '2016']
### Next, we iterate over all the columns to find any numbers > 1
for col in year_columns:
    bool_vector = elections_flattened[col] > 1
    print('For ' + str(col) + ' the number of voters that had more than one voting record is ' + str(sum(bool_vector)))

First, let's count how many votes were cast in each year counting multiple votes

for col in year_columns:
    print('For ' + str(col) + ' the number of votes cast were ' + str(int(sum(elections_flattened[col]))))

Now we need to change all of these numbers > 1 to 1

for col in year_columns:
    elections_flattened[col] = elections_flattened[col] >= 1

Let's once again look at how many votes were cast.  Odd years should be the same as before, and even years should be less.

for col in year_columns:
    print('For ' + str(col) + ' the number of votes cast were ' + str(int(sum(elections_flattened[col]))))

Now we need to pull in the demographics files for 2016, 2015, and 2015 to calculate turnout for each of these years

## First we do 2014

local.database = 'registration2014'
local.table = 'VR_Snapshot_20141104'

I want all rows with status codes 'A' and 'I'

condition1 = "[status_cd] = 'A'"
condition2 = "[status_cd] = 'I'"
condition = "(" + condition1 + " OR " + condition2 + ")"  
print condition

#Set up the fields that I need to pull from the database 
fields = '[county_desc],[ncid],[precinct_abbrv]'
print fields

#Pull the data from the database
demographics=pullODBC(local.driver,
                      local.server,
                      local.database,
                      local.username,
                      local.password,
                      local.table,
                      fields,
                      condition)

print(len(demographics))
print demographics.tail(n=15)
raw_len = len(demographics)# Save the number of rows

We now need to delete any rows with blank county_desc, ncid, or precinct_abbrv.  We do that by first changing blans to NaNs, and the deleting any rows with NaNs

def fix_blanks_NaN(x):
    if x == "":
        return np.nan
    elif x == " ":
        return np.nan
    else:
        return x

#Change blank fields to Nan
fields = demographics.columns
for col in fields:
    demographics[col] = demographics[col].apply(fix_blanks_NaN)
#Next, delete any rows with NaN and disply counts
demographics.dropna(inplace=True, how='any') #drop all rows with NaN
clean_len=len(demographics) #how many rows to I have left?
difference = raw_len - clean_len #how many rows have I dropped?
fraction = float(difference)/float(raw_len) #what fraction of rows have I dropped?
print "The number of rows I have dropped is " + str(difference) 
print "The number of rows remaining in the database is " + str(clean_len)
print "The precentage of rows that I have dropped is: " + str('{:.3%}'.format(fraction))

Now we do a left join of elections_flattened to demographics since we use the 2014 registration data as our denominator for 2014 turnout

combined_2014 = pd.merge(demographics, elections_flattened, how='left', on='ncid')
print(len(combined_2014), len(demographics), len(elections_flattened))# combined should have the same number of rows as demographics

Let's take a look at combined_2014

print(combined_2014.tail(n=15))

We now need to fill all the NaNs with False.  The only NaNs are for registered voters without voting records that were created by the left join

combined_2014.fillna(value=False, method=None, axis=None, inplace=True, limit=None, downcast=None)

print(combined_2014.tail(n=15))

Next, we create a new field that combines county_desc with precinct_abbrv

combined_2014['county_pct'] = combined_2014['county_desc']+combined_2014['precinct_abbrv']

print(combined_2014.head(n=15))

Subset the dataframe to keep only the needed columns

turnout = combined_2014.ix[:,['county_desc',  'county_pct', '2014']]

Convert '2014' to 1s and 0s

turnout['2014'] = turnout['2014'].astype(int)

print(turnout.head())

Strip out blanks in 'county_pct'

turnout['county_pct'] = turnout['county_pct'].map(lambda x: x.replace(" ", ""))
print(turnout.head())

Use pandas pivot table functionality to create a table with turnout by count_precinct combination

First, we create a custome aggfunc to yield the fraction voted

def fraction_one(x):
    frac = np.sum(x)/np.size(x)
    return(frac)

Create a Pivot Table with index of county_precinct and column and aggregation '2014'

turnout_2014 = pd.pivot_table(turnout, index=['county_pct'], values=['2014'],  
                                 aggfunc=fraction_one,  margins = False) 

print len(turnout_2014)
print(turnout_2014.head())

Convert to a dataframe

turnout_2014 = pd.DataFrame.from_records(turnout_2014.to_records())
print turnout_2014.head()

## Now we do the same thing for 2015

local.database = 'registration2015'
local.table = 'VR_Snapshot_20151103'

I want all rows with status codes 'A' and 'I'

condition1 = "[status_cd] = 'A'"
condition2 = "[status_cd] = 'I'"
condition = "(" + condition1 + " OR " + condition2 + ")"  
print condition

#Set up the fields that I need to pull from the database 
fields = '[county_desc],[ncid],[precinct_abbrv]'
print fields

#Pull the data from the database
demographics=pullODBC(local.driver,
                      local.server,
                      local.database,
                      local.username,
                      local.password,
                      local.table,
                      fields,
                      condition)

print(len(demographics))
print demographics.tail(n=15)
raw_len = len(demographics)# Save the number of rows

We now need to delete any rows with blank county_desc, ncid, or precinct_abbrv.  We do that by first changing blans to NaNs, and the deleting any rows with NaNs

def fix_blanks_NaN(x):
    if x == "":
        return np.nan
    elif x == " ":
        return np.nan
    else:
        return x

#Change blank fields to Nan
fields = demographics.columns
for col in fields:
    demographics[col] = demographics[col].apply(fix_blanks_NaN)
#Next, delete any rows with NaN and disply counts
demographics.dropna(inplace=True, how='any') #drop all rows with NaN
clean_len=len(demographics) #how many rows to I have left?
difference = raw_len - clean_len #how many rows have I dropped?
fraction = float(difference)/float(raw_len) #what fraction of rows have I dropped?
print "The number of rows I have dropped is " + str(difference) 
print "The number of rows remaining in the database is " + str(clean_len)
print "The precentage of rows that I have dropped is: " + str('{:.3%}'.format(fraction))

Now we do a left join of elections_flattened to demographics since we use the 2015 registration data as our denominator for 2015 turnout

combined_2015 = pd.merge(demographics, elections_flattened, how='left', on='ncid')
print(len(combined_2015), len(demographics), len(elections_flattened))# combined should have the same number of rows as demographics

Let's take a look at combined_2015

print(combined_2015.tail(n=15))

We now need to fill all the NaNs with False.  The only NaNs are for registered voters without voting records that were created by the left join

combined_2015.fillna(value=False, method=None, axis=None, inplace=True, limit=None, downcast=None)

print(combined_2015.tail(n=15))

Next, we create a new field that combines county_desc with precinct_abbrv

combined_2015['county_pct'] = combined_2015['county_desc']+combined_2015['precinct_abbrv']

print(combined_2015.head(n=15))

Subset the dataframe to keep only the needed columns

turnout = combined_2015.ix[:,['county_desc',  'county_pct', '2015']]

Convert '2015' to 1s and 0s

turnout['2015'] = turnout['2015'].astype(int)

print(turnout.head())

Strip out blanks in 'county_pct'

turnout['county_pct'] = turnout['county_pct'].map(lambda x: x.replace(" ", ""))
print(turnout.head())

Use pandas pivot table functionality to create a table with turnout by count_precinct combination

Create a Pivot Table with index of county_precinct and column and aggregation '2015'

turnout_2015 = pd.pivot_table(turnout, index=['county_pct'], values=['2015'],  
                                 aggfunc=fraction_one,  margins = False) # Use the aggfunc created for 2014

print len(turnout_2015)
print(turnout_2015.head())

Convert to a dataframe

turnout_2015 = pd.DataFrame.from_records(turnout_2015.to_records())
print turnout_2015.head()

## Now we do the same thing for 2016

local.database = 'registration2016'
local.table = 'VR_Snapshot_20161108'

I want all rows with status codes 'A' and 'I'

condition1 = "[status_cd] = 'A'"
condition2 = "[status_cd] = 'I'"
condition = "(" + condition1 + " OR " + condition2 + ")"  
print condition

#Set up the fields that I need to pull from the database 
fields = '[county_desc],[ncid],[precinct_abbrv]'
print fields

#Pull the data from the database
demographics=pullODBC(local.driver,
                      local.server,
                      local.database,
                      local.username,
                      local.password,
                      local.table,
                      fields,
                      condition)

print(len(demographics))
print demographics.tail(n=15)
raw_len = len(demographics)# Save the number of rows

We now need to delete any rows with blank county_desc, ncid, or precinct_abbrv.  We do that by first changing blans to NaNs, and the deleting any rows with NaNs

def fix_blanks_NaN(x):
    if x == "":
        return np.nan
    elif x == " ":
        return np.nan
    else:
        return x

#Change blank fields to Nan
fields = demographics.columns
for col in fields:
    demographics[col] = demographics[col].apply(fix_blanks_NaN)
#Next, delete any rows with NaN and disply counts
demographics.dropna(inplace=True, how='any') #drop all rows with NaN
clean_len=len(demographics) #how many rows to I have left?
difference = raw_len - clean_len #how many rows have I dropped?
fraction = float(difference)/float(raw_len) #what fraction of rows have I dropped?
print "The number of rows I have dropped is " + str(difference) 
print "The number of rows remaining in the database is " + str(clean_len)
print "The precentage of rows that I have dropped is: " + str('{:.3%}'.format(fraction))

Now we do a left join of elections_flattened to demographics since we use the 2016 registration data as our denominator for 2016 turnout

combined_2016 = pd.merge(demographics, elections_flattened, how='left', on='ncid')
print(len(combined_2016), len(demographics), len(elections_flattened))# combined should have the same number of rows as demographics

Let's take a look at combined_2016

print(combined_2016.tail(n=15))

We now need to fill all the NaNs with False.  The only NaNs are for registered voters without voting records that were created by the left join

combined_2016.fillna(value=False, method=None, axis=None, inplace=True, limit=None, downcast=None)

print(combined_2016.tail(n=15))

Next, we create a new field that combines county_desc with precinct_abbrv

combined_2016['county_pct'] = combined_2016['county_desc']+combined_2016['precinct_abbrv']

print(combined_2016.head(n=15))

Subset the dataframe to keep only the needed columns

turnout = combined_2016.ix[:,['county_desc',  'county_pct', '2016']]

Convert '2016' to 1s and 0s

turnout['2016'] = turnout['2016'].astype(int)

print(turnout.head())

Strip out blanks in 'county_pct'

turnout['county_pct'] = turnout['county_pct'].map(lambda x: x.replace(" ", ""))
print(turnout.head())

Use pandas pivot table functionality to create a table with turnout by count_precinct combination

Create a Pivot Table with index of county_precinct and column and aggregation '2016'

turnout_2016 = pd.pivot_table(turnout, index=['county_pct'], values=['2016'],  
                                 aggfunc=fraction_one,  margins = False) # Use the aggfunc created for 2014

print len(turnout_2016)
print(turnout_2016.head())

Convert to a dataframe

turnout_2016 = pd.DataFrame.from_records(turnout_2016.to_records())
print turnout_2016.head()

## Now we will use outer joins to combine the 3 dataframes

turnout_all = pd.merge(turnout_2016, turnout_2015, how='outer', on='county_pct')
print(len(turnout_all), len(turnout_2016), len(turnout_2015))# turnout_all  should have more rows than 2016 and 2015
turnout_all = pd.merge(turnout_all, turnout_2014, how='outer', on='county_pct')
print(len(turnout_all), len(turnout_2014))# the new turnout_all should have more rows than the old one and 2014

Now we write this to a csv for analysis and inclusion as 3 dimensions in precinct clustering

turnout_all.to_csv('turnout_all.csv')
