==> code/9h - Calculating Vote Fractions for Clustering.ipynb <==
##  We're going calculate vote fraction for Trump and Newton by county/precinct combination to use as 2 dimensions for our clustering work

#First set everthing up
import pyodbc
import graphlab as gl
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import datetime
import config #server information and credentials
from pullODBC import pullODBC
from pullODBCall import pullODBCall
from matplotlib.backends.backend_pdf import PdfPages
from __future__ import division # convert integer division to floating point
%matplotlib inline
plt.style.use('seaborn-white')
plt.rcParams['font.size'] = 7
plt.tight_layout;
plt.rc('xtick', labelsize=7) 
plt.rc('ytick', labelsize=7) 

#Set up the fields that I need to pull from the database 
fields = 'candidate, contest_name, election_date, precinct, total_votes, county'
print fields

#Set up the condition
condition0 = "contest_name = 'NC ATTORNEY GENERAL'"
condition1 = "contest_name = 'US PRESIDENT'"
condition2 = "election_date='2016-11-08'"
condition = "(" + condition0 + " OR " + condition1 + ") AND " + condition2
print condition

Read in the results for President and Attorney General for 2016 General Election

#Pull the data from the database
results=pullODBC(config.driver,
                      config.server,
                      config.database,
                      config.username,
                      config.password,
                      config.table,
                      fields,
                      condition)

print(len(results))
print results.tail(n=15)
raw_len = len(results)# Save the number of rows

We now need to clean the data to remove any rows with blank columns

def fix_blanks_NaN(x):
    if x == "":
        return np.nan
    elif x == " ":
        return np.nan
    else:
        return x

#Change blank fields to Nan
fields = results.columns
for col in fields:
    results[col] = results[col].apply(fix_blanks_NaN)
#Next, delete any rows with NaN and disply counts
results.dropna(inplace=True, how='any') #drop all rows with NaN
clean_len=len(results) #how many rows to I have left?
difference = raw_len - clean_len #how many rows have I dropped?
fraction = float(difference)/float(raw_len) #what fraction of rows have I dropped?
print "The number of rows I have dropped is " + str(difference) 
print "The number of rows remaining in the database is " + str(clean_len)
print "The precentage of rows that I have dropped is: " + str('{:.3%}'.format(fraction))

The data is clean.  No rows have been dropped

Now we create a now column county_pct that combines county and precinct and strip out any blanks in this new column.  

results['county_pct'] = results['county']+results['precinct'] # create new column
results['county_pct'] = results['county_pct'].map(lambda x: x.replace(" ", "")) # strip out the blanks
print(results.head(n=15))


Now we create a pivot table to summarize the results

results_pivot = pd.pivot_table(results, index=['county_pct'], values=['total_votes'],  
                                 aggfunc=np.sum,  columns = ['contest_name','candidate'], margins = False) 
results_pivot.head()

We now need to fill all the NaNs with 0  

results_pivot.fillna(value=0, method=None, axis=None, inplace=True, limit=None, downcast=None)
results_pivot.head()

Now we add new columns that sum across contest_name

results_pivot[['contest_total_ag', 'contest_total_president']] = results_pivot['total_votes'].sum(axis=1, level=0)
results_pivot.head()

results_pivot['fraction_newton'] = results_pivot['total_votes', 'NC ATTORNEY GENERAL','Buck Newton']/results_pivot['contest_total_ag']
results_pivot['fraction_trump'] = results_pivot['total_votes', 'US PRESIDENT','Donald J. Trump']/results_pivot['contest_total_president']

Subset the dataframe to keep only the needed columns

results_pivot.head()

Create clean dataframe to convert to csv

results_pivot['county_pct'] = results_pivot.index
results_df = results_pivot[['county_pct','fraction_newton','fraction_trump']]
results_df.to_csv('results.csv')


print results_df.head()
