==> code/9i -  Cleaning, Standardizing, and Joining Data for Clustering.ipynb <==
##  We're going collect and clean all the data to create a final Data Frame for Clustering

#First set everthing up
import pyodbc
import graphlab as gl
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import datetime
import config #server information and credentials
from pullODBC import pullODBC
from pullODBCall import pullODBCall
from matplotlib.backends.backend_pdf import PdfPages
from __future__ import division # convert integer division to floating point
%matplotlib inline
plt.style.use('seaborn-white')
plt.rcParams['font.size'] = 7
plt.tight_layout;
plt.rc('xtick', labelsize=7) 
plt.rc('ytick', labelsize=7) 

We create a utility function that will standardize all columns except the first one (which should contain a key) to mean 0 and standard deviation of 1

def standardize(df):
    means = df.mean()
    std = df.std()
    size = len(df.columns)
    df.iloc[:,1:size] = (df.iloc[:,1:size]- means)/std
    return df
    

Next, we read the file containing election results by county/precinct combination, add a column 'trump_premium' that contains the fraction voted for Trump minus fraction voted for Newton, standardize this, and inspect the resulting data frame to ensure correctness.

results = pd.read_csv('results_V3.csv')
#results = results.iloc[:, 0:4]
results['trump_premium'] = results['fraction_trump'] - results['fraction_newton']
results = standardize(results)

results.head()

len(results)

print ('The means are ' + str(results.mean()))
print ('The standard deviations are ' + str(results.std()))

Next, we read the file precinct demographics by county/precinct combination, standardize this, and inspect the resulting data frame to ensure correctness.

demographics = pd.read_csv('precinct_demographics_nov2.csv')
#demographics = demographics.iloc[:, 0:]
demographics = standardize(demographics)

demographics.head()

len(demographics)

print ('The means are ' + str(demographics.mean()))
print ('The standard deviations are ' + str(demographics.std()))

Now we do an outer join and write to csv to find errors

combined_pct = pd.merge(demographics, results, how='outer')
combined_pct.to_csv('combined_pct.csv')

After observing and fixing errors, we now do a left join with demographic on the left and save this in a csv

combined_pct = pd.merge(demographics, results, how='left')
combined_pct.to_csv('combined_pct.csv')

Next, we read the file containing election turnout by county/precinct combination, standardize this, and inspect the resulting data frame to ensure correctness.

turnout = pd.read_csv('turnout_allV3.csv')
#turnout = turnout.iloc[:, 0:]
turnout = standardize(turnout)

turnout.head()

len(turnout)

print ('The means are ' + str(turnout.mean()))
print ('The standard deviations are ' + str(turnout.std()))

We need to strip out all the blanks in column 'county_pct'

turnout['county_pct'] = turnout['county_pct'].map(lambda x: x.strip())
combined_pct['county_pct'] = combined_pct['county_pct'].map(lambda x: x.strip())

Now we do an outer join and write to csv to find errors

combined_pct_test = pd.merge(combined_pct, turnout, how='outer')
combined_pct_test.to_csv('combined_pct_test.csv')

After observing and fixing errors, we now replace all NaNs with 0 (mean) and do a left join with demographic on the left and save this in a csv

combined_pct_final = pd.merge(combined_pct, turnout, how='left')
combined_pct_final.fillna(value=0, inplace=True)
combined_pct_final.to_csv('combined_pct_final.csv')
